Streamlit-Apps/neural-style-transfer/requirements.txt
streamlit==1.23.1
numpy==1.23.5
tensorflow-hub==0.13.0
tensorflow==2.12.0
Pillow==9.5.0

